// packages/@ros-cdk/services/edas/lib/edas.generated.ts
export interface RosClusterProps {

    /**
     * @Property clusterName: Cluster name
     */
    readonly clusterName: string;

    /**
     * @Property clusterType: Cluster type. 1-Swarm cluster, 2-ECS cluster, 3-Kubernetes Cluster
     */
    readonly clusterType: number;

    /**
     * @Property networkMode: Network Type. 1- classic network, 2-VPC
     */
    readonly networkMode: number;

    /**
     * @Property logicalRegionId: Custom namespace RegionId (format: Physical Region: custom namespace identifier)
     */
    readonly logicalRegionId?: string;

    /**
     * @Property oversoldFactor: Docker CPU cluster oversold. Support 2 (1: 2 ratio) / 4 (1: 4) / 8 (1: 8 ratio)
     */
    readonly oversoldFactor?: number;

    /**
     * @Property vpcId: VPC network ID. If network selection VPC, this parameter Required
     */
    readonly vpcId?: string;
}
// ...
export class RosCluster extends ros.RosResource {
    /**
     * The resource type name for this resource class.
     */
    public static readonly ROS_RESOURCE_TYPE_NAME = "ALIYUN::EDAS::Cluster";

    /**
     * A factory method that creates a new instance of this class from an object
     * containing the properties of this ROS resource.
     */

    /**
     * @Attribute ClusterId: Cluster ID
     */
    public readonly attrClusterId: any;

    /**
     * @Attribute ClusterName: Cluster name
     */
    public readonly attrClusterName: any;

    /**
     * @Attribute ClusterType: Cluster type
     */
    public readonly attrClusterType: any;

    /**
     * @Attribute IaasProvider: Provider
     */
    public readonly attrIaasProvider: any;

    public enableResourcePropertyConstraint: boolean;


    /**
     * @Property clusterName: Cluster name
     */
    public clusterName: string;

    /**
     * @Property clusterType: Cluster type. 1-Swarm cluster, 2-ECS cluster, 3-Kubernetes Cluster
     */
    public clusterType: number;

    /**
     * @Property networkMode: Network Type. 1- classic network, 2-VPC
     */
    public networkMode: number;

    /**
     * @Property logicalRegionId: Custom namespace RegionId (format: Physical Region: custom namespace identifier)
     */
    public logicalRegionId: string | undefined;

    /**
     * @Property oversoldFactor: Docker CPU cluster oversold. Support 2 (1: 2 ratio) / 4 (1: 4) / 8 (1: 8 ratio)
     */
    public oversoldFactor: number | undefined;

    /**
     * @Property vpcId: VPC network ID. If network selection VPC, this parameter Required
     */
    public vpcId: string | undefined;

    /**
     * Create a new `ALIYUN::EDAS::Cluster`.
     *
     * @param scope - scope in which this resource is defined
     * @param id    - scoped id of the resource
     * @param props - resource properties
     */
    constructor(scope: ros.Construct, id: string, props: RosClusterProps, enableResourcePropertyConstraint: boolean) {
        super(scope, id, { type: RosCluster.ROS_RESOURCE_TYPE_NAME, properties: props });
        this.attrClusterId = ros.Token.asString(this.getAtt('ClusterId'));
        this.attrClusterName = ros.Token.asString(this.getAtt('ClusterName'));
        this.attrClusterType = ros.Token.asString(this.getAtt('ClusterType'));
        this.attrIaasProvider = ros.Token.asString(this.getAtt('IaasProvider'));

        this.enableResourcePropertyConstraint = enableResourcePropertyConstraint;
        this.clusterName = props.clusterName;
        this.clusterType = props.clusterType;
        this.networkMode = props.networkMode;
        this.logicalRegionId = props.logicalRegionId;
        this.oversoldFactor = props.oversoldFactor;
        this.vpcId = props.vpcId;
    }


    protected get rosProperties(): { [key: string]: any }  {
        return {
            clusterName: this.clusterName,
            clusterType: this.clusterType,
            networkMode: this.networkMode,
            logicalRegionId: this.logicalRegionId,
            oversoldFactor: this.oversoldFactor,
            vpcId: this.vpcId,
        };
    }
    protected renderProperties(props: {[key: string]: any}): { [key: string]: any }  {
        return rosClusterPropsToRosTemplate(props, this.enableResourcePropertyConstraint);
    }
}